+++ Areas/Account/Data/IServices/AP/IAPDocSetOffService.cs
@@ -6,7 +6,7 @@ namespace AEMSWEB.Areas.Account.Data.IServices.AP
 {
     public interface IAPDocSetOffService
     {
-        public Task<APDocSetOffViewModelCount> GetAPDocSetOffListAsync(short CompanyId, short pageSize, short pageNumber, string searchString, string fromDate, string toDate, short UserId);
+        public Task<APDocSetOffViewModelCount> GetAPDocSetOffListAsync(short CompanyId, int pageSize, int pageNumber, string searchString, string fromDate, string toDate, short UserId);
 
         public Task<APDocSetOffViewModel> GetAPDocSetOffByIdAsync(short CompanyId, long SetoffId, string SetoffNo, short UserId);
