+++ Areas/Account/Data/IServices/AP/IAPRefundService.cs
@@ -6,7 +6,7 @@ namespace AEMSWEB.Areas.Account.Data.IServices.AP
 {
     public interface IAPRefundService
     {
-        public Task<APRefundViewModelCount> GetAPRefundListAsync(short CompanyId, short pageSize, short pageNumber, string searchString, string fromDate, string toDate, short UserId);
+        public Task<APRefundViewModelCount> GetAPRefundListAsync(short CompanyId, int pageSize, int pageNumber, string searchString, string fromDate, string toDate, short UserId);
 
         public Task<APRefundViewModel> GetAPRefundByIdAsync(short CompanyId, long RefundId, string RefundNo, short UserId);
