+++ Areas/Account/Data/Services/AP/APCreditNoteService.cs
@@ -27,7 +27,7 @@ namespace AEMSWEB.Areas.Account.Data.Services.AP
             _accountService = accountService;
         }
 
-        public async Task<APCreditNoteViewModelCount> GetAPCreditNoteListAsync(short CompanyId, short pageSize, short pageNumber, string searchString, string fromDate, string toDate, short UserId)
+        public async Task<APCreditNoteViewModelCount> GetAPCreditNoteListAsync(short CompanyId, int pageSize, int pageNumber, string searchString, string fromDate, string toDate, short UserId)
         {
             APCreditNoteViewModelCount countViewModel = new APCreditNoteViewModelCount();
             try
