+++ Areas/Account/Data/Services/AP/APDebitNoteService.cs
@@ -27,7 +27,7 @@ namespace AEMSWEB.Areas.Account.Data.Services.AP
             _accountService = accountService;
         }
 
-        public async Task<APDebitNoteViewModelCount> GetAPDebitNoteListAsync(short CompanyId, short pageSize, short pageNumber, string searchString, string fromDate, string toDate, short UserId)
+        public async Task<APDebitNoteViewModelCount> GetAPDebitNoteListAsync(short CompanyId, int pageSize, int pageNumber, string searchString, string fromDate, string toDate, short UserId)
         {
             APDebitNoteViewModelCount countViewModel = new APDebitNoteViewModelCount();
             try
