+++ Areas/Account/Data/Services/AP/APDocSetOffService.cs
@@ -27,7 +27,7 @@ namespace AEMSWEB.Areas.Account.Data.Services.AP
             _accountService = accountService;
         }
 
-        public async Task<APDocSetOffViewModelCount> GetAPDocSetOffListAsync(short CompanyId, short pageSize, short pageNumber, string searchString, string fromDate, string toDate, short UserId)
+        public async Task<APDocSetOffViewModelCount> GetAPDocSetOffListAsync(short CompanyId, int pageSize, int pageNumber, string searchString, string fromDate, string toDate, short UserId)
         {
             APDocSetOffViewModelCount countViewModel = new APDocSetOffViewModelCount();
             try
