+++ Areas/Account/Data/Services/AP/APRefundService.cs
@@ -27,7 +27,7 @@ namespace AEMSWEB.Areas.Account.Data.Services.AP
             _accountService = accountService;
         }
 
-        public async Task<APRefundViewModelCount> GetAPRefundListAsync(short CompanyId, short pageSize, short pageNumber, string searchString, string fromDate, string toDate, short UserId)
+        public async Task<APRefundViewModelCount> GetAPRefundListAsync(short CompanyId, int pageSize, int pageNumber, string searchString, string fromDate, string toDate, short UserId)
         {
             APRefundViewModelCount countViewModel = new APRefundViewModelCount();
             try
