+++ Areas/Account/Data/Services/AR/ARAdjustmentService.cs
@@ -27,7 +27,7 @@ namespace AEMSWEB.Areas.Account.Data.Services.AR
             _accountService = accountService;
         }
 
-        public async Task<ARAdjustmentViewModelCount> GetARAdjustmentListAsync(short CompanyId, short pageSize, short pageNumber, string searchString, string fromDate, string toDate, short UserId)
+        public async Task<ARAdjustmentViewModelCount> GetARAdjustmentListAsync(short CompanyId, int pageSize, int pageNumber, string searchString, string fromDate, string toDate, short UserId)
         {
             ARAdjustmentViewModelCount countViewModel = new ARAdjustmentViewModelCount();
             try
