+++ Areas/Account/Data/Services/AR/ARDocSetOffService.cs
@@ -27,7 +27,7 @@ namespace AEMSWEB.Areas.Account.Data.Services.AR
             _accountService = accountService;
         }
 
-        public async Task<ARDocSetOffViewModelCount> GetARDocSetOffListAsync(short CompanyId, short pageSize, short pageNumber, string searchString, string fromDate, string toDate, short UserId)
+        public async Task<ARDocSetOffViewModelCount> GetARDocSetOffListAsync(short CompanyId, int pageSize, int pageNumber, string searchString, string fromDate, string toDate, short UserId)
         {
             ARDocSetOffViewModelCount countViewModel = new ARDocSetOffViewModelCount();
             try
