+++ Areas/Account/Data/Services/CB/CBBankTransferService.cs
@@ -27,7 +27,7 @@ namespace AEMSWEB.Areas.Account.Data.Services.CB
             _accountService = accountService;
         }
 
-        public async Task<CBBankTransferViewModelList> GetCBBankTransferListAsync(short CompanyId, short pageSize, short pageNumber, string searchString, short UserId)
+        public async Task<CBBankTransferViewModelList> GetCBBankTransferListAsync(short CompanyId, int pageSize, int pageNumber, string searchString, short UserId)
         {
             CBBankTransferViewModelList countViewModel = new CBBankTransferViewModelList();
             try
