+++ Areas/Account/Data/Services/CB/CBGenPaymentService.cs
@@ -27,7 +27,7 @@ namespace AEMSWEB.Areas.Account.Data.Services.CB
             _accountService = accountService;
         }
 
-        public async Task<CBGenPaymentViewModel> GetCBGenPaymentListAsync(short CompanyId, short pageSize, short pageNumber, string searchString, string fromDate, string toDate, short UserId)
+        public async Task<CBGenPaymentViewModel> GetCBGenPaymentListAsync(short CompanyId, int pageSize, int pageNumber, string searchString, string fromDate, string toDate, short UserId)
         {
             CBGenPaymentViewModel countViewModel = new CBGenPaymentViewModel();
             try
