+++ Areas/Admin/Data/IAllLogService.cs
@@ -6,12 +6,12 @@ namespace AEMSWEB.Areas.Admin.Data
 {
     public interface IAllLogService
     {
-        public Task<IEnumerable<AuditLogViewModel>> GetAuditLogListAsync(short CompanyId, short pageSize, short pageNumber, string searchString, short UserId);
+        public Task<IEnumerable<AuditLogViewModel>> GetAuditLogListAsync(short CompanyId, int pageSize, int pageNumber, string searchString, short UserId);
 
-        public Task<UserLogViewModelCount> GetUserLogListAsync(short CompanyId, short pageSize, short pageNumber, string searchString, short UserId);
+        public Task<UserLogViewModelCount> GetUserLogListAsync(short CompanyId, int pageSize, int pageNumber, string searchString, short UserId);
 
         public Task<SqlResponse> SaveUserLog(short CompanyId, AdmUserLog admUserLog, short UserId);
 
-        public Task<IEnumerable<ErrorLogViewModel>> GetErrorLogListAsync(short CompanyId, short pageSize, short pageNumber, string searchString, short UserId);
+        public Task<IEnumerable<ErrorLogViewModel>> GetErrorLogListAsync(short CompanyId, int pageSize, int pageNumber, string searchString, short UserId);
     }
 }
